Split up GpuOwnerOverlapCopy to have one class per file
